requirements.txt
# FunASR 及其依赖
funasr>=1.0.0
modelscope>=1.9.0
addict
# PyTorch (根据 CUDA 版本选择)
torch>=2.0.0
torchaudio>=2.0.0

# 音频处理
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# 文本处理
cn2an>=0.5.0  # 中文数字转换
pypinyin>=0.49.0  # 拼音处理
jieba>=0.42.0  # 中文分词

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 评估指标
editdistance>=0.6.0  # 编辑距离计算
tqdm>=4.65.0  # 进度条

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 类型检查
typing-extensions>=4.7.0
